perf(nvim): start ~200ms faster from within lazygit
Typically I want to use lazygit inside of neovim as an embedded floating
application. In these cases I sometimes want to navigate to a file that
lazygit displays as having changed.

This starts neovim, but it used to be slow/sluggish to start. This
is because it was running the normal startup sequence, which includes
checking for updates, loading plugins, etc.

Now I load a minimal set of plugins and settings when starting from
lazygit. This makes the startup time much faster.