feat(nvim): make snacks `lines` picker not hide the text
Changing the default layout seems to override something for the `lines`
picker. It was difficult to fix this properly because something seems to
be inherited inside of snacks when the default layout is set, so I did
something arduous and stupidly simple instead.

Might be related to https://github.com/folke/snacks.nvim/issues/788