feat(mika): group fixup_commits in branch-summary (#406)
**Issue:**

When working on a branch for a long time, I tend to create many fixup
commits. These look quite confusing in the branch-summary output.

**Solution:**

Group the fixup commits under their base commit, so that the
branch-summary output is cleaner and easier to read.

I tend to create two kinds of these:

1. > fixup! <commit message>
2. > fixup! <commit message>
   >
   > additional explanation for the fixup