fix: fixups not being handled in `mika summary`
Recently I have loved the idea of using `mika branch-summary` to
generate PR review notes from a stack of commits.

However, I noticed that when `mika summary` is used, the fixup commits
are not included in the parent commits' messages hierarchically, making
it very difficult to read through the changes.

Implement handling of fixup commits in the summary generation to fix
this.